Fixed `crypto:sha_mac/2 undef` error
`erlang:function_exported/3` returns `false` if a module has not yet been loaded. When running Erlang in development mode (e.g. in IntelliJ/VSCode) modules are loaded lazily (and therefore the crypto module may not yet be loaded at this point), causing this function to incorrectly detect an OTP<22 version of `crypto`.